Configure ApolloClient to not cache SSR fetch results
It is unclear how `revalidatePath` works, since we could not get it to work after making changes according to the next.js documentation. It appears that next.js caches the result of `fetch` made by the `ApolloClient` which makes the pages to render the old data.  Hence we set `fetchOptions: { cache: "no-store" }` to prevent that caching.

We will revisit this once next.js `revalidatePath` works as expected (there seems to be several bugs reported).